+++ src/index.html
@@ -6,7 +6,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
   <title>Paseito</title>
-  <link href="style.css" rel="stylesheet" type="text/css">
+  <link href="./style/style.css"rel="stylesheet" type="text/css">
   <link rel="preconnect" href="https://fonts.googleapis.com">
   <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
 <link href="https://fonts.googleapis.com/css2?family=Cairo:wght@400;700&display=swap" rel="stylesheet">
@@ -16,7 +16,7 @@
 <body>
 
   <div id="root"></div> <!--contenedor de vistas-->
-  <script type="module" src="main.js"></script>
+  <script type="module" src="./routes/main.js"></script>
