+++ src/lib/firestoreConfig.js
@@ -1,0 +1,15 @@
+import { getFirestore, addDoc, collection, getDocs, query, where, deleteDoc, doc, setDoc, getDoc, updateDoc } from 'https://www.gstatic.com/firebasejs/9.6.7/firebase-firestore.js';
+
+export {
+  getFirestore,
+  addDoc,
+  collection,
+  getDocs,
+  query,
+  where,
+  deleteDoc,
+  doc,
+  setDoc,
+  getDoc,
+  updateDoc,
+};
