+++ src/views/start.js
@@ -1,5 +1,5 @@
 // eslint-disable-next-line import/no-cycle
-import { onNavigate } from './main.js';
+import { onNavigate } from '../routes/main.js'; //eslint-disable-line
 
 function start() {
   const startButtons = document.createElement('div');
